pyproject.toml
[tool.poetry]
name = "ff14-dataset"
version = "0.1.0"
description = "FFXIV dataset builder for ML: event-driven + tick state, Parquet/DuckDB, PySide6 GUI"
authors = ["you <you@example.com>"]
readme = "README.md"
packages = [{ include = "ff14_dataset", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
polars = "^1.6.0"
duckdb = "^1.0.0"
httpx = { version = "^0.27.0", extras = ["http2"] }
orjson = "^3.10.0"
pydantic = "^2.7.0"
python-dotenv = "^1.0.1"
PySide6 = ">=6.8.0.2,<7.0.0"
prefect = "^3.0.0"
tenacity = "^9.0.0"
PyYAML = "^6.0.2"
tqdm = "^4.66.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.5.0"
types-PyYAML = "^6.0.12.20240808"

[tool.poetry.scripts]
ff14ds-gui = "ff14_dataset.gui.app:main"
ff14ds-cli = "ff14_dataset.cli:main"

[tool.ruff]
line-length = 100
